app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
coverage==7.6.1
dnspython==2.6.1
email_validator==2.2.0
Faker==28.1.0
fastapi==0.112.3
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
h11==0.14.0
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.31.5
MarkupSafe==2.1.5
motor==3.5.1
msgpack==1.0.8
packaging==24.1
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
pytest==8.3.2
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
requests==2.32.3
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.0.3
